### 2025-02-27 #176 @fcdimitr
Hi @sloede 👋 ,

would you like to review this JuliaCon proceedings submission?

you can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).
