### 2025-06-17 #176 @brianguenter
I looked at the paper again. It's too far outside my expertise for me to give an informed review.
