### 2025-08-27 #176 @mongibellili
Thanks for the invitation, but I’ll have to decline — I’m quite busy at the moment and the paper is outside my main area of expertise.
