### 2025-09-16 #176 @CasBex
Thanks for the effort in looking for reviewers @fcdimitr . I talked to Tobias Thummerer (@ThummeTo) at the Modelica and FMI conference last week and he agreed to review if this would be ok with you?
