### 2025-09-17 #176 @fcdimitr
Hi @ThummeTo 👋 ,

would you like to review this JuliaCon proceedings submission?

you can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).
