### 2025-09-29 #187 @editorialbot
[REVIEW]: FMIExchange.jl: Foreign Models Through a Familiar Interface
**Submitting author:** @CasBex (<a href="http://orcid.org/0000-0003-0614-7225">Lucas Bex</a>)
**Repository:** https://github.com/Electa-Git/FMIExchange.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.0
