### 2025-09-29 #187 @fcdimitr
Hello all. You can ignore the `Failed to discover a Statement of need section in paper` warning. We do not explicitly enforce this in JuliaCon submissions.

Reviewers, please generate your checklist by typing `@editorialbot generate my checklist`.
