### 2025-10-29 #187 @ThummeTo
Should the review (of the paper) be posted here directly? Like an open review?
