### 2025-11-12 #187 @CasBex
Judging by some other review threads, open review is the intention.
- https://github.com/JuliaCon/proceedings-review/issues/181
- https://github.com/JuliaCon/proceedings-review/issues/124
