### 2025-11-21 #176 @ThummeTo
Just double-checking, current version of the paper is [this](https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00176/jcon.00176/10.21105.jcon.00176.pdf)?
