### 2025-11-21 Electa-Git/FMIExchange.jl#13 @ThummeTo
Tests fail on Windows
Hi,

running Windows 11 and Julia 1.11.6, I find the following when doing a `test FMIExchange` in a fresh environment:

```
ERROR: LoadError: AssertionError: You are trying to build FMUs but `omc` is not in your PATH
Stacktrace:
 [1] top-level scope
   @ C:\Users\thummeto\.julia\packages\FMIExchange\2u8Ui\deps\build.jl:43
 [2] include(fname::String)
   @ Main .\sysimg.jl:38
 [3] top-level scope
   @ C:\Users\thummeto\.julia\packages\FMIExchange\2u8Ui\test\runtests.jl:11
 [4] include(fname::String)
   @ Main .\sysimg.jl:38
 [5] top-level scope
   @ none:6
in expression starting at C:\Users\thummeto\.julia\packages\FMIExchange\2u8Ui\deps\build.jl:40
in expression starting at C:\Users\thummeto\.julia\packages\FMIExchange\2u8Ui\test\runtests.jl:8
ERROR: Package FMIExchange errored during testing
```

Relevant for [the review](https://github.com/JuliaCon/proceedings-review/issues/187).

Thanks and best!
