### 2025-11-21 #187 @ThummeTo
W.r.t. the "functionality" checkbox of the review checklist see [this issue](https://github.com/Electa-Git/FMIExchange.jl/issues/13) I just opened.
