### 2025-11-21 Electa-Git/FMIExchange.jl#13 @CasBex
Hi ThummeTo thanks for the issue. This seems to be related to the build process of the FMUs, which requires OpenModelica and relevant libraries installed on your machine (see also the README). If you just want to run the tests you can also manually compile the FMUs with whatever Modelica compiler you have available and then run the tests that way.

To avert the problem of building the FMUs I created a docker image in the deps folder, but it seems that this one is currently broken as well. We manually compile OMC so we don't depend on the OpenModelica versions available in the OpenModelica repositories (old versions tend to get removed relatively quickly). However, this manual compilation process seems to be broken. I will look into that issue next week, but it may take a bit longer to solve.
