+++ Project.toml
@@ -1,28 +1,31 @@
 name = "FMIExchange"
 uuid = "d0f58030-e4f1-4214-a600-bc055fa6d38e"
-authors = ["Cas Bex <cas.bex@kuleuven.be>"]
 version = "0.1.0"
+authors = ["Cas Bex <cas.bex@kuleuven.be>"]
 
 [deps]
 DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
+FMIBase = "900ee838-d029-460e-b485-d98a826ceef2"
 FMICore = "8af89139-c281-408e-bce2-3005eb87462f"
 FMIImport = "9fcbc62e-52a0-44e9-a616-1359a0008194"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 
 [compat]
-DiffEqCallbacks = "2"
-FMICore = "0.20"
-FMIImport = "0.16.4"
+DiffEqCallbacks = "2,3,4"
+FMIBase = "1"
+FMICore = "1"
+FMIImport = "1"
 Requires = "1"
 SciMLBase = "2"
 julia = "1.6"
 
 [extras]
+ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["DelimitedFiles", "Interpolations", "OrdinaryDiffEq", "Test"]
+test = ["DelimitedFiles", "Interpolations", "OrdinaryDiffEq", "Test", "ADTypes"]
