### 2025-11-25 aa1e9f1
Update OMC version and use new bouncing ball model for tests
