+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"Defining_and_Simulating_Models_short/#Defining-and-Simulating-Models","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"","category":"section"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"This guide explains the mathematical formalism on which this package is based, and how that translates to package usage.","category":"page"},{"location":"Defining_and_Simulating_Models_short/#Mathematical-Background","page":"Defining and Simulating Models","title":"Mathematical Background","text":"","category":"section"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"In the context of this package, a model is a system ordinary differential equations with events (further hybrid ODE). A model has inputs, states and outputs: inputs are external to the model; states evolve according to some hybrid ODE; outputs are an arbitrary function of time, inputs and states.","category":"page"},{"location":"Defining_and_Simulating_Models_short/#Working-with-Model-Exchange-FMUs","page":"Defining and Simulating Models","title":"Working with Model-Exchange FMUs","text":"","category":"section"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"A Model-Exchange FMU contains a system described by a differential algebraic equation with events (hybrid DAE), but exposes this to the user as a hybrid ODE by internally resolving all algebraic relations. The FMU can be used to compute derivatives, outputs and event indicators. Performing integration and triggering events at the appropriate time is the task of the external solver. Event handlers inside the FMU are activated when the external solver triggers an event. These handlers change the internal FMU state.","category":"page"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"FMIExchange.jl was developed to make the bridge between the FMU and the external solver, which is DifferentialEquations.jl. The main idea is to provide functionality to automatically convert the FMU object as imported by FMIImport.jl to a callable function that can be used by DifferentialEquations.jl and to automatically generate the callbacks to trigger FMU events.","category":"page"},{"location":"Defining_and_Simulating_Models_short/#Composing-Simulations-With-FMIExchange.jl","page":"Defining and Simulating Models","title":"Composing Simulations With FMIExchange.jl","text":"","category":"section"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"To incorporate FMUs in a DifferentialEquations.jl workflow, we need to fit it into a function of one of the forms du = f(u, p, t) or f!(du, u, p, t). To do this FMIExchange.jl takes the following approach:","category":"page"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"All states are stored in u\nAll inputs and outputs are stored in p","category":"page"},{"location":"Defining_and_Simulating_Models_short/","page":"Defining and Simulating Models","title":"Defining and Simulating Models","text":"As such, u and p act as memory buffers for the models, and should be some subtype of AbstractVector. If a simulation comprises multiple models, their memory buffers should not overlap to avoid
[…]
