### 2025-11-26 e48e745
Incorporate comments ThummeTo
