+++ paper/paper.tex
@@ -10,7 +10,7 @@
 \hypersetup{colorlinks}
 
 \usepackage{tikz}
-\usetikzlibrary{positioning, arrows.meta, fit}
+\usetikzlibrary{positioning, arrows.meta, fit, calc}
 
 \begin{document}
 
@@ -19,50 +19,52 @@
 \maketitle
 
 \begin{abstract}
-
 The Functional Mock-up Interface (FMI) standard \cite{modelicaassociation} facilitates interoperability of differential algebraic equation (DAE) simulation software.
-Compliant DAE models (so-called Functional Mock-up Units (FMUs)) can be simulated through a standardized interface, which allows to easily exchange these models between different software tools.
+Compliant models (so-called Functional Mock-up Units (FMUs)) can be simulated through a standardized interface, which allows to easily exchange these models between different software tools.
 The Julia ecosystem contains an extensive FMI package suite that supports (among others) simulation, import, export and differentiation of FMUs.
-This paper presents FMIExchange.jl, which complements existing FMI packages by interfacing FMUs with the popular DifferentialEquations.jl package \cite{rackauckas2017}.
-Additionally, FMIExchange.jl provides tools to compose simulations containing multiple FMUs and native Julia models.
-
+This paper presents FMIExchange.jl, which provides utilities to simplify interfacing FMUs with the popular DifferentialEquations.jl package \cite{rackauckas2017}.
+Additionally, FMIExchange.jl provides tools to compose simulations combining multiple FMUs and native Julia models.
 \end{abstract}
 
 \section{Statement of Need}
-An FMU (\cref{fig:FMU}) transforms a DAE with events\footnote{Events here signify discontinuities in the solution: time-based events are discontinuities at fixed times, typically implemented through discrete callbacks in DifferentialEquations.jl; state-based events depend on the integrated quantities and are typically implemented with continuous callbacks.} \cref{eq:DAE} into an ordinary differential equation (ODE) with events \cref{eq:ODE}.
+The Functional Mock-up Interface (FMI) standard \cite{modelicaassociation} presents a standardised software interface for simulating an ordinary differential equation (ODE) with events to enhance interoperability of simulation and modelling tools.
+Many differential algebraic equation (DAE) modelling tools have the possibility to export models (after transformation to an ODE) to Functional Mock-Up Units (FMU, \cref{fig:FMU}), which comply with this interface.
 
-In addition to the states $u$ and inputs $p_i$ required for simulating the model, the FMU can define arbitrary outputs $p_o$ \cref{eq:output} that are exposed to the user.
-Two types of FMU are available: Co-Simulation FMUs only expose inputs and outputs and come bundled with an integrator and event handler, Model Exchange (ME-)FMUs require an external ODE integrator/event handler and expose all quantities; this text focuses on ME-FMUs.
+FMUs define a set of states states $u$ and inputs $p_i$ required for simulating the model \cref{eq:ODE}, as well as a set of outputs $p_o$ \cref{eq:output} which are exposed to the user.
+FMUs may contain events, that cause discontinuouties in the solution of the ODE.
+Events may trigger at fixed times or whenever the state/inputs fulfil some condition.
[…]
