+++ paper/ref.bib
@@ -12,6 +12,15 @@
   doi = {10.5334/jors.151}
 }
 
+@online{bex2024,
+  title = {{{FMIExchange demo}}},
+  author = {Bex, Lucas},
+  year={2024},
+  month={Jan},
+  day={19},
+  howpublished = {url: \url{https://gitlab.kuleuven.be/u0138844/fmiexchangedemo}}
+}
+
 @online{modelicaassociation,
   title = {{{Functional Mock-up Interface}}},
   author = {Modelica Association},
