+++ paper/paper.yml
@@ -4,6 +4,7 @@ keywords:
   - Differential equation
   - Functional Mock-up Interface
   - Simulation
+  - Model Integration
 authors:
   - name: Lucas Bex
     orcid: 0000-0003-0614-7225
