### 2025-11-27 #187 @CasBex
Thank you @ThummeTo for the review. I have updated the paper with your comments (https://github.com/Electa-Git/FMIExchange.jl/commit/e48e7452c98b80a9d6aa32c61d35dc3bc161b148) and present my rebuttal inline below. With respect to the functionality issue, I will keep the discussion in the issue thread you created.

> FMIExchange.jl provides functionality for combining native Julia ODEs with FMUs. This is done by providing a more accessible interface for the underlying Julia structs that correspond to the ME-FMU (such as ODEProblem) than existing libraries (e.g. FMI.jl) do. This is complemented by further optimizations, like caching of signals and implementation of fast addressing maps. The general topic is relevant and the presented library can be helpful for model integration tasks („integration“ in terms of combination). At some points (see below) the contribution needs to be described more precisely, it is sometimes hard to distinguish from related work. A deeper motivation and examples would further highlight the relevance of the work.
> # Major
>
>  In the first sentence „An FMU (Fig. 1) transforms a DAE with events (1) into an ordinary differential equation (ODE) with events (2)“. This is incorrect. An FMU can be seen as a model container and is not capable of performing a transformation (index reduction is described here). Index reduction (DAE to ODE) is performed by the model compiler (e.g., Modelica compiler) during compilation before the export.

The original sentence was intended to show that FMU export is used a lot in DAE modelling software. The wording has been changed to show that this modelling software exports DAEs to ODEs and then compiles those into an FMU.

>  I would propose to sharpen the difference between what is not possible in existing libraries (like FMI.jl), but explicitly added in FMIExchange.jl. Some sentences are worded vaguely. For example, in the abstract (and again later in Sec. 3) “FMIExchange.jl […] which complements existing FMI packages by interfacing FMUs with the popular DifferentialEquations.jl package” leads to the assumption that existing FMI packages are not able to embed ME-FMUs into DifferentialEquations.jl and DiffEqCallbacks.jl – but this is exactly how simulation of ME-FMUs is implemented e.g. in FMI.jl. If the interface of FMIExchange.jl allows for more freedom, this could be illustrated clearly, for example, with a compact code example (FMIExchange.jl on the left, FMI.jl on the right), or alternatively by showcasing what is possible (“enabled”) with FMIExchange.jl. For my understanding, the key benefit is the simplified model integration (“integration” in terms of “combination”) of multiple FMU(s) and Julia models, together with additional features (e.g. fast address maps if multiple models are involved). It might make sense to show a short code example, combining a ME-FMU and a Julia-function.

In the abstract wording has been changed from "interfacing FMUs with the popular DifferentialEquations.jl package" to "provides utilities to simplify interfacing FMUs with the popular DifferentialEquations package". In section 3, it has been explicitly noted that while it is possible to simulate multiple FMUs and native ODEs without FMIExchange.jl, the main contribution of it is the simplification through its interface. Due to space constraints (max 2 pages) a code example could not be included: FMIExchange.jl is mostly useful for more complex simulations with multiple components, so a tiny example would not provide much benefits. Instead we refer in the text to a demo available online for interested readers.

> # Minor
>
>  In Fig. 2: Events and Event-handling is (technically) not part of the ODE-solver, but the simulation environment, that again controls the ODE-solver for the continuous solution sections, monitors the event indicators, etc.

Solver in figure 2 was renamed to Simulation Environment.

> Eq. (1) and (2) mention the term “events”, without mathematical introduction. I propose to clarify that the DAE (1) and ODE (2) formalisms are shown without events in the equations, and mention that it is a simplification (neglecting events) in the text. Adding the term “events” alone to the mathematical description does not offer added value, IMO.

The word events was removed from equations (1), (2). Instead events were described in-text.

> In Sec. 2, there are two sentences containing three dots “backs/dynamics functions. . . to use in the DifferentialEquations.jl” and “dynamics function. . .”. I am not sure if this is intended.

Dots were removed and where applicable wording was altered.

> Sec. 1 states that there are two types of FMUs, technically there are three (scheduled execution was added in FMI3).

The third type of FMU was added to the description of the types.

> Fig. 1 may lead to the assumption, that for CS-FMUs the solver is not part of the FMU (it is labeled “internal solver”, however is outside the “FMU” rectangle), which it is. Further, for CS, the event indicators are not exposed. Maybe just showing the ME-FMU in Fig. 1 will be more clear.

The CS-FMU was included within the FMU box and the caption was expanded to describe both FMU types. If the figure remains unclear the CS-FMU box may be removed from the figure entirely.

> # Further opportunities for optimization
>
> Composing models (in terms of merging state vectors and the right-hand sides) is often referred to as “model integration” in literature (maybe use as keyword to increase range).

An additional keyword was added.

> A motivation with areas of application would further elaborate on the necessity of the approach.

We think the application areas of FMUs are sufficiently wide as to not warrant a single recommendation for an application domain. The demo we link in Section 2 is a demo for demand response in energy systems, which serves as a motivating example.

> A small example with code and results would further underscore the relevance of the paper and (maybe) motivate the reader to implement an own application.

As mentioned above, a demo was added.

> Some scientists don't like it when only two papers are cited. I don't see this as a problem in a very software-focused paper, but I wanted to point it out briefly. However, I do not want to encourage people to quote excessively for the sake of just having a lot of quotes. Naming applications could lead to some "natural" citations here.

We agree with the reviewer that the limited bibliography is not a significant problem and decide to leave it at the current amount of bibliographical references.
