### 2025-11-27 Electa-Git/FMIExchange.jl#13 @CasBex
Hi @ThummeTo I've updated the package to use the newer FMI.jl versions (major version 1) and use more recent Openmodelica versions in the docker image. However, for running the tests on your local machine I haven't found a solution except for including precompiled FMUs for both windows and Linux in the repository itself. I do not have access to a license to export FMUs without runtime license using commercial Modelica software (e.g. Dymola), and as far as I'm aware, cross-compilation on Openmodelica currently doesn't work very well. If you want to verify the functionality of FMIExchange.jl I see some possibilities:

- Rely on CI, which compiles the FMUs in a Linux container and runs the tests in a Linux container
- Compile the FMUs on your host machine either manually or through the build script. Instructions in the README are available on how to do this, and the `deps` folder contains a list of Modelica dependencies for the models which are all open source. If anything is unclear, let me know and I can update the instructions accordingly.
- Download the FMUs from the CI artifacts and run tests on a compatible Linux machine
- Test only the reference bouncing ball FMU, which to my knowledge is compiled for both Linux and Windows. This would mean you need to skip the build step and comment out part of the tests.

I'd like to stress that the issues presented here are a matter of compiling the FMU for the correct platform with the correct dependencies, and are in no way inherent to FMIExchange.jl. Any other FMI software also will not be able to run an FMU which was not compiled for the correct target
