### 2025-12-02 Electa-Git/FMIExchange.jl#13 @ThummeTo
So checking the script and CI output is fine for me. (can be closed)

Just in case you are interested, we solved this with [FMIZoo.jl](https://github.com/ThummeTo/FMIZoo.jl), so adding a test dependency to a repo containing binary FMUs.
