### 2025-12-02 #187 @ThummeTo
Thanks for the reply and for taking the comments into consideration.

# Paper

The page limit of only two pages is, of course, somewhat restrictive. I understand that in this context, the sample code cannot really be placed in a meaningful way (I was thinking of a code block like [this](https://github.com/Electa-Git/FMIExchange.jl/blob/b19a35a6c8fb75f321cf80b8bf7b485312140e8b/test/bouncingball.jl#L12)).

It's great that you've added an example notebook that shows how to use the library. This gives you a quick overview of the range of functions.

W.r.t. the written contribution, I don't have further objections.

# Library

As you can see above, three checkboxes are still unchecked. "Methodology" and "Results" can be argued by being presented in the linked notebook instead of the paper. "Community guidelines" can be fixed by adding the corresponding phrases to the readme.md in the repository root. With this very minor remaining adjustment, I approve it on my side.
