### 2025-12-08 #187 @CasBex
Hi @ThummeTo thank you again for taking the time for this. If I understand your comment correctly the [pull request above](https://github.com/Electa-Git/FMIExchange.jl/pull/15) should be sufficient and I do not need to add any new things for the Methodology and Results section right?
