### 2026-01-20 #187 @CasBex
Dear editor @fcdimitr

If we cannot reach @H-Sax for the review, might we look for an alternative reviewer? I could query the [BOPTEST mailing list](https://groups.google.com/g/ibpsa-boptest/about) for interested parties with the correct expertise who might be willing to review instead. This group has long used FMUs to simulate building models and as such I believe that we can find suitable reviewers with no conflict of interest.

I am of course open to other ideas for completing this review process. However, as it has now been 18 months since JuliaCon 2024, I would like to ensure that the review can move forward without further delay.

Please let me know your thoughts.
