### 2026-01-27 #187 @CasBex
I have no contacts in the FMI design group, and I did not get answers from the IBPSA BOPTEST mailing list. However, one of the professors in my department may be a suitable reviewer @hakanergun .

Prof. Ergun specialises in power system optimisation and is co-author of packages like PowerModelsACDC.jl . He is affiliated with Electa, ESAT, KU Leuven and Energyville (which is also my affiliation). He is otherwise not involved in my work as PhD researcher or in any projects that I'm involved in. If you consider this conflict of interest too great I completely understand, but if that is the case I would need some help from the editor to find a good reviewer.
