### 2026-02-09 #187 @fcdimitr
Hi @hakanergun!

Would you like to review this JuliaCon proceedings submission?

You can read more about reviewing for JuliaCon proceedings in the [reviewer guide](https://juliacon.github.io/proceedings-guide/reviewer/).

Regarding COI: If you can confirm that you were not involved in the development of the work (e.g., providing detailed feedback on drafts, helping with analysis, or co-developing ideas), it's fine to proceed as a reviewer.
