### 2026-02-11 #187 @hakanergun
- [x] I confirm that I have read the [JuliaCon conflict of interest policy](https://juliacon.github.io/proceedings-guide/reviewer/#conflict_of_interest) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JCon for the purpose of this review.

- **Comment:** Generally yes, however the restriction on the XLSX package might be too strict.

- [x] **Authors:** Does the `paper.pdf` file include a list of authors with their affiliations?
- [x] **A statement of need:** Does the paper clearly explain the problem it addresses, its importance to the broader community, the intended audience, and how it connects to existing work in the field?
- **Comment:** Generally, yes. For improvment: The introduction part could have mentioned the keyword "co-simulation". The developed package can be very useful for optimisation models in dynamic systems where after a steady-state optimisation based on algebraic equations, a dynamic simulation can be triggered relying on ODEs. Providing some examples of such use cases would make the paper much stronger and more accessible to a broader community that might not be too familiar with FMIs.
