### 2026-02-16 #187 @CasBex
Thank you for the review @hakanergun. Find below my rebuttal to your comments.

> * **Comment:** Generally yes, however the restriction on the XLSX package might be too strict.*

FMIExchange depends only on SciMLBase, DiffEqCallbacks, packages from the FMI ecosystem and Requires.jl . Normally there should be no dependency on XLSX. I think the dependency requirements are pretty loose as well. Is it possible that there was a dependency conflict with something else in your main environment? If the problem is on the side of FMIExchange.jl I'll be glad to help out if you make an issue in [the repository](https://github.com/Electa-Git/FMIExchange.jl).

> * **Comment:** Generally, yes. For improvment: The introduction part could have mentioned the keyword "co-simulation". The developed package can be very useful for optimisation models in dynamic systems where after a steady-state optimisation based on algebraic equations, a dynamic simulation can be triggered relying on ODEs. Providing some examples of such use cases would make the paper much stronger and more accessible to a broader community that might not be too familiar with FMIs.

Co-simulation is now mention in the statement of need. W.r.t. the optimisation problem: the provided demo implements a model predictive controller for an energy system, which fits your suggested use case. To make this more clear, I gave the demo description its own short paragraph with more details. To respect the 2-page constraint, minor textual adjustments were made elsewhere in the text.

Please confirm whether these changes suffice.
