+++ paper/paper.tex
@@ -99,7 +99,7 @@ To avoid having to manually assign addresses and provide a human-friendly interf
 The demo \cite{bex2024} provides a practical example of a FMIExchange.jl worklflow.
 The demo implements a model-predictive (MPC) and a rule-based controller for a household energy system.
 In this demo, there is some model mismatch between the optimisation model of the MPC and the physical system.
-The effect of this model mismatch is through dynamic simulation of the system.
+The effect of this model mismatch is studied through dynamic simulation of the system.
 
 \section{FMU Utilities}
 Function Mock-Up Units are merely a special kind of model to include in a simulation.
