### 2026-02-16 #187 @CasBex
New pdf generation due to erratum in the text.
