### 2026-02-19 #187 @CasBex
Thanks to both reviewers for their time and effort.
@fcdimitr what are the next steps for finalising the publication?
