### 2026-02-19 #187 @fcdimitr
Thank you all for your help!

@CasBex Could you please check the Zenodo archive that is required for acceptance?
You need to follow [the instructions here](https://juliacon.github.io/proceedings-guide/author/#publication_phase) so that there is a permanent DOI for the final (accepted) version of the code & paper.

Then, we can accept it and close this issue. Thanks!
