### 2026-02-20 247fd03
Add FMIExchange.jl DOI and zenodo reference
