+++ paper/paper.tex
@@ -11,6 +11,7 @@
 
 \usepackage{tikz}
 \usetikzlibrary{positioning, arrows.meta, fit, calc}
+\newcommand{\typename}[1]{\textit{#1}}
 
 \begin{document}
 
@@ -22,7 +23,7 @@
 The Functional Mock-up Interface (FMI) standard \cite{modelicaassociation} facilitates interoperability of differential algebraic equation (DAE) simulation software.
 Compliant models (so-called Functional Mock-up Units (FMUs)) can be simulated through a standardized interface, which allows to easily exchange these models between different software tools.
 The Julia ecosystem contains an extensive FMI package suite that supports (among others) simulation, import, export and differentiation of FMUs.
-This paper presents FMIExchange.jl, which provides utilities to simplify interfacing FMUs with the popular DifferentialEquations.jl package \cite{rackauckas2017}.
+This paper presents FMIExchange.jl \cite{bex2025e}, which provides utilities to simplify interfacing FMUs with the popular DifferentialEquations.jl package \cite{rackauckas2017}.
 Additionally, FMIExchange.jl provides tools to compose simulations combining multiple FMUs and native Julia models.
 \end{abstract}
 
@@ -42,7 +43,7 @@ g(u, p_{i}, t) &= p_o \label{eq:output}
 
 \begin{figure}
   \center
-\begin{tikzpicture}[node distance=0.5cm, >={Stealth[round]}]
+\begin{tikzpicture}[node distance=0.5cm, >={Stealth[round]}, scale=1.0, transform shape]
   \node (FMUeq) [draw=black] {$\begin{aligned}
     \dot{u} &= f(u, p_{i}, t) \\
     p_o &= g(u, p_i, t)
@@ -53,11 +54,11 @@ g(u, p_{i}, t) &= p_o \label{eq:output}
 \node (y) [right=of FMU] {$p_o$} edge [<-] (FMU);
 % \node (xdot) [below=of FMU, xshift=0.75cm] {$\dot{u}$} edge [<-] (FMU.south -| xdot);
 % \node (x) [below=of FMU, xshift=-0.75cm] {$u$} edge [->] (FMU.south -| x);
-\node (ME) [below=1cm of FMUeq, draw=black, label={[name=MEL]below:\footnotesize Model Exchange}] {External Solver};
+\node (ME) [below=0.75cm of FMUeq, draw=black, label={[name=MEL]below:\footnotesize Model Exchange}] {External Solver};
 \coordinate (e1) at ($(FMUeq.south east)!0.5!(FMUeq.south)$ |- ME.north);
 \coordinate (e2) at ($(FMUeq.south west)!0.5!(FMUeq.south)$ |- ME.north);
-\draw [->] (e1) -- node [auto, swap, yshift=-1mm] {$\dot{u}$} node [auto, yshift=-1mm] {\footnotesize \shortstack{event\\indicators}} ($(ME.north west)!(e1)!(ME.north east)$);
-\draw [<-] (e2) -- node [auto, yshift=-1mm] {$u$} node [auto, swap, yshift=-1mm] {\footnotesize \shortstack{event\\triggers}} ($(ME.north west)!(e2)!(ME.north east)$);
+\draw [->] (e1) -- node [auto, swap, yshift=-1mm] {$\dot{u}$} node [auto, yshift=-1mm] {\footnotesize \shortstack{event indicators}} ($(ME.north west)!(e1)!(ME.north east)$);
+\draw [<-] (e2) -- node [auto, yshift=-1mm] {$u$} node [auto, swap, yshift=-1mm] {\footnotesize \shortstack{event triggers}} ($(ME.north west)!(e2)!(ME.north east)$);
[…]
