+++ paper/ref.bib
@@ -57,3 +57,15 @@
   keywords = {FMI,FMU,Julia,NeuralFMU,NeuralODE},
   file = {/home/cas/Zotero/storage/YQHZJG9W/Thummerer et al. - 2021 - NeuralFMU Towards Structural Integration of FMUs .pdf}
 }
+@software{bex2025e,
+  title = {{{FMIExchange}}.jl},
+  author = {Bex, Lucas and Deconinck, Geert},
+  date = {2025-11-26},
+  doi = {10.5281/zenodo.18710589},
+  url = {https://zenodo.org/records/18710589},
+  urldate = {2026-02-20},
+  abstract = {FMIExchange.jl is a Julia package that provides ways to easily load Model Exchange Functional Mock-up Units (FMUs) and simulate them using the DifferentialEquations.jl package.},
+  langid = {english},
+  organization = {Zenodo},
+  keywords = {Differential Equations,Functional Mock-up Interface,Julia,Model Integration,Simulation},
+}
