### 2026-02-20 #187 @CasBex
I've generated the DOI for version 0.2.0 of the code (which is the code after review), available here: https://doi.org/10.5281/zenodo.18710590

In the paper I've cited FMIExchange.jl using the DOI to the most recent version of FMIExchange.jl: https://doi.org/10.5281/zenodo.18710589 which at the time of writing resolves to 0.2.0 but when a new update is published, it will refer to the newer version.

Due to space constraints minor modifications were made to the formatting: FMIExchange.jl type names were italicized (rather than monospaced), and a minor adjustment has been made to the layout of figure 1.

Let me know if this works for you.
