### 2026-02-26 #187 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
